Create compress-image.yml
This will be used every time there is a push to the main branch. The job used is compress-images and uses the latest Ubuntu. It does the following steps: 1) checks out the repository code (actions/checkout), 2) installs the sharp package, and 3) compresses the images in the image folder of the repository. Quality - sets JPEG and PNG compression to whatever you like. Resize and Format are also adjustable.